Added the particleFX plugin as a requirement. Updated the appImage script to be more explanitory.

+++ linuxBuild/buildAppImage.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -x
 
-#CMAKE_BUILD_TYPE="Debug"
+#Run this command with your pwd as the output build directory of the engine.
 
 START_DIR="$1"
 if [ -v ${1} ]; then
@@ -8,7 +8,11 @@ if [ -v ${1} ]; then
     exit 1
 fi
 
-#cd ${1}
+#Check we're in the right directory.
+if [ ! -f av ]; then
+    echo "Missing engine executable in pwd."
+    exit 1
+fi
 
 rm -rf outAppDir
 mkdir -p outAppDir/usr/bin
@@ -17,5 +21,5 @@ cp -r Hlms essential avSetup.cfg outAppDir/usr/bin
 #You have to add the extract-and-run while in the container
 linuxdeploy --appimage-extract-and-run \
     --appdir outAppDir/ \
-    -e av -e RenderSystem_GL3Plus.so \
+    -e av -e RenderSystem_GL3Plus.so -e Plugin_ParticleFX.so \
     -i ${1}/setup/logo.svg -d ${1}/setup/entry.desktop --output appimage
